requirements.txt
python-telegram-bot==20.0
typing-extensions
exceptiongroup
database
